game: spell the phrase step words as long long
The step words have an eight-byte stride (sll 3 at 0x001b91d8) and are
read and written with ld and sd by PlayBar, ApplyStepMask,
PostRemixFxMsg, and PostBarStatusMsg, so PhraseDatabase, PhraseMgr,
and JamEffectsMgr now use long long, and Save narrows explicitly to the
long overload that writes the low word.

+++ include/game/jameffectsmgr.h
@@ -69,7 +69,7 @@ public:
      * @param nMask The step's mask, one bit per effect type.
      * @ghidraAddress 0x001a56d8
      */
-    void ApplyStepMask(long nMask);
+    void ApplyStepMask(long long nMask);
 
 private:
     // For a message on this track, flips the effect's bit in the step value of the message's bar,
